moved index into src directory
