+++ src/tests/indexSpec.ts
@@ -1,4 +1,4 @@
-import { myFunc } from '../../index'
+import { myFunc } from '../index'
 
 it('expect myFunc(5) to equal 25', () => {
     expect(myFunc(5)).toEqual(25)
